# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Workload Evaluation API
  version: 0.1.0
  description: >
    本地工作量评估系统 API。响应统一为 { code, message, data, requestId? }。

    **鉴权（以运行代码为准）**：当前实现使用 JWT，请求头 `Authorization: Bearer <token>`。
    本文档中仍出现的 `X-Role` 参数为历史草稿，**运行时不再作为权限依据**。

    **未在本文档中展开的已实现端点**（完整列表见仓库
    `03_技术设计/系统演进/实现与文档对齐说明.md`）包括但不限于：
    `/api/v1/auth/*`、`/api/v1/versions/*`、`/api/v1/system/*`、
    `/api/v1/ai/parse-basic-info`、`/api/v1/ai/company-profile-summary`、`/api/v1/ai/kimi-assessment/preview`、`/api/v1/ai/kimi-assessment/export-markdown`、
    `/api/v1/teams/*`、`GET /api/v1/wbs`；
    下载路径为 `GET /downloads/{fileName}`（非 `/api/v1` 前缀）。

    **`/api/v1/agent/*` 说明**：当前已挂载 V1 最小基座 `POST /api/v1/agent/chat`。
    该接口返回统一 JSON 事件数组；抽取、PDF、真实归档工具和前端工作台仍在后续批次接线。
servers:
  - url: http://localhost:3000
    description: Local API server
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/harness/runs:
    post:
      tags: [Harness]
      summary: Create Harness run
      operationId: createHarnessRun
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                mode:
                  type: string
                  enum: [interactive, replay, regression]
                aiSessionId:
                  type: string
      responses:
        "200":
          description: Harness run created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericOkResponse"
    get:
      tags: [Harness]
      summary: List current user's Harness runs
      operationId: listHarnessRuns
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Harness run list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericOkResponse"
...
  /api/v1/harness/runs/{runId}/files:
    post:
      tags: [Harness]
      summary: Bind a file metadata to a Harness run
      operationId: bindHarnessFile
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: runId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [attachmentId, fileName]
              properties:
                attachmentId:
                  type: string
                fileName:
                  type: string
                fileSize:
                  type: integer
                mimeType:
                  type: string
                fileHash:
                  type: string
                storagePath:
                  type: string
                role:
                  type: string
                roleConfidence:
                  type: number
      responses:
        "200":
          description: File bound
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericOkResponse"
  /api/v1/harness/runs/{runId}/parse-result:
    post:
      tags: [Harness]
      summary: Submit parsed file evidence into Harness
      operationId: submitHarnessParseResult
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: runId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sourceFile]
              properties:
                fileId:
                  type: string
                sourceFile:
                  type: string
                sheets:
                  type: array
                  items:
                    type: string
                summary:
                  type: object
                  properties:
                    projectName:
                      type: string
                    customerName:
                      type: string
                    industry:
                      type: string
                items:
                  type: array
                  items:
                    type: object
                    required: [text]
                    properties:
                      sourceSheet:
                        type: string
                      sourceCell:
                        type: string
                      category:
                        type: string
                      text:
                        type: string
                      metadata:
                        type: object
      responses:
        "200":
          description: Parsed evidence accepted; returns Harness run detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericOkResponse"
        "400":
          description: Invalid Harness stage for parse result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Harness run not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/harness/runs/{runId}/report-v1:
    post:
      tags: [Harness]
      summary: Generate Harness requirement report v1 through the configured model
      operationId: generateHarnessRequirementReportV1
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: runId
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                force:
                  type: boolean
      responses:
        "200":
          description: Requirement report v1 generated; returns Harness run detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericOkResponse"
        "400":
          description: Model call failed, schema validation failed, or run is in an invalid stage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Harness run not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/harness/runs/{runId}/report-v2:
    post:
      tags: [Harness]
      summary: Generate Harness requirement report v2 from evidence + v1 + answers
      operationId: generateHarnessRequirementReportV2
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: runId
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                force:
                  type: boolean
      responses:
        "200":
          description: Requirement report v2 generated; returns Harness run detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericOkResponse"
        "400":
          description: Model call failed, schema validation failed, missing v1 report, or run is in an invalid stage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Harness run not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/harness/runs/{runId}/answers:
    post:
      tags: [Harness]
      summary: Submit structured answers to a Harness run
      operationId: submitHarnessAnswers
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: runId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                answers:
                  type: array
                  items:
                    type: object
                    properties:
                      field:
                        type: string
                      value: {}
                      source:
                        type: string
                        enum: [user_chat, structured_form]
      responses:
        "200":
          description: Answers submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericOkResponse"
  /api/v1/harness/runs/{runId}/actions/{actionId}/confirm:
    post:
      tags: [Harness]
      summary: Confirm or cancel a Harness action
      operationId: confirmHarnessAction
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: runId
          required: true
          schema:
            type: string
        - in: path
          name: actionId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                confirmed:
                  type: boolean
                actionType:
                  type: string
                  description: Harness confirmable action type. Known confirmed actions require the run to have reached `report_v2_ready`. `enter_formal_estimation` additionally requires the URL `actionId` to equal `enter_formal_estimation`, and creates draft project/assessment records for manual review; it does not publish a formal estimate. Repeating `enter_formal_estimation` for the same `(runId, actionType)`, or another known action for the same `(runId, actionId, actionType)`, is idempotent and returns the existing confirmed ToolEvent.
      responses:
        "200":
          description: Action confirmation recorded. For `enter_formal_estimation`, `data.event.output.project` and `data.event.output.assessmentDraft` contain the created draft links and `data.run.metadata.links.assessmentVersionId` tracks the assessment draft. If draft creation fails, a failed ToolEvent is persisted with errorMessage and the API returns an error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericOkResponse"
  /api/v1/harness/runs/{runId}/retry:
    post:
      tags: [Harness]
      summary: Retry a failed Harness run
      operationId: retryHarnessRun
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: runId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Run retry started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericOkResponse"
  /api/v1/harness/runs/{runId}/reanalyze:
    post:
      tags: [Harness]
      summary: Force reanalysis of a Harness run
      operationId: reanalyzeHarnessRun
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: runId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reanalysis flag set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericOkResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "登录 `POST /api/v1/auth/login` 获取 token，请求头 `Authorization: Bearer <token>`"
...
  schemas:
    GenericOkResponse:
      type: object
      properties:
        code:
          type: integer
          const: 0
        message:
          type: string
          const: ok
        requestId:
          type: string
        data:
          type: object
          additionalProperties: true
      required: [code, message, data]
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        requestId:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              reason:
                type: string
            required: [field, reason]
      required: [code, message, details, requestId]
...